+++ api/auth/handler.ts
@@ -1,50 +1,6 @@
-const jwt = require('jsonwebtoken');
 import { Profile } from './profile';
 
 const profile = new Profile();
-
-const AUTH0_CLIENT_ID = 'hfDx6WXS2nkcLUhOcHe0Xq34lZE3wfrH';
-const AUTH0_CLIENT_SECRET = 'wvSHEEB3V_VvnuwxIDWSqukWoI3tTcqf28YYpKndZEXn3pYj3Q0ueJTDpR6ZT_B8';
-
-// Policy helper function
-function generatePolicy(principalId, effect, resource) {
-    const authResponse: any = {};
-    authResponse.principalId = principalId;
-    if (effect && resource) {
-        const policyDocument: any = {};
-        policyDocument.Version = '2012-10-17';
-        policyDocument.Statement = [];
-        const statementOne: any = {};
-        statementOne.Action = 'execute-api:Invoke';
-        statementOne.Effect = effect;
-        statementOne.Resource = resource;
-        policyDocument.Statement[0] = statementOne;
-        authResponse.policyDocument = policyDocument;
-    }
-    return authResponse;
-}
-
-// Reusable Authorizer function, set on `authorizer` field in serverless.yml
-export function auth(event, context, cb) {
-    if (event.authorizationToken) {
-        // remove "bearer " from token
-        console.log('event', JSON.stringify(event));
-        const token = event.authorizationToken.substring(7);
-        const options = {
-            audience: AUTH0_CLIENT_ID,
-        };
-        const secretBuffer = new Buffer(AUTH0_CLIENT_SECRET, 'base64');
-        jwt.verify(token, secretBuffer, options, (err, decoded) => {
-            if (err) {
-                cb('[401] Unauthorized');
-            } else {
-                cb(null, generatePolicy(decoded.sub, 'Allow', event.methodArn));
-            }
-        });
-    } else {
-        cb('[401] Unauthorized');
-    }
-}
 
 export function getAllProfiles(event, context, callback) {
     profile.getAll()
