+++ package.json
@@ -5,9 +5,16 @@
   "main": "handler.js",
   "scripts": {
     "start": "sls deploy",
-    "test": "nyc mocha test/*"
+    "test": "nyc mocha test/*",
+    "serverless-offline": "sls offline start"
   },
   "nyc": {
+    "watermarks": {
+      "lines": [80, 95],
+      "functions": [80, 95],
+      "branches": [80, 95],
+      "statements": [80, 95]
+    },
     "extension": [
       ".ts",
       ".tsx"
@@ -19,7 +26,8 @@
     ],
     "require": [
       "ts-node/register"
-    ]
+    ],
+    "source-map": true
   },
   "author": "",
   "license": "ISC",
